Always add new high score if there's room
Originally, only two conditions were checked before adding a new high
score: was this list empty, or was the current score better than the
worst of the old scores. This missed one obvious case: is there any room
on the high score list, regardless of how good the current score is.