Set clicked class unconditionally in handler
When the clicked/flagged/bomb class cleanup was made, setting the class
to 'clicked' was made conditional on the poly not being a bomb. This
broke some things, and the class changes appear to have not worked 100%
correctly anyhow.

This commit makes the clicked class uncoditional again, and uses CSS
selectors to make it look right.