Remove window.onresize callback
Calling updateSize() for every resize event worked great on desktop
browsers (where it was rarely needed), but caused numerous problems
on mobile browsers (where zooming is necessary for accuracy).

The "solution" is to just set the window size once, when the board is
being built, and use that for the rest of the game. If the user wishes
to update the board size again, they'll have to start a new game.