Break svg box shadow to fix board size
The svg element width should now be determined by #theBoard, instead of
whatever the browser decides to set it to. This sets the right and left
fuzzy edges of the box outside of the page, but that's better than
having a game board that's a fraction of the size that it should be.

This problem was originally noticed in Chrome, while playing around with
extra block elements in the layout.